Run pdftotext and OCR on PDF county files
Create and cache .txt.gz versions of the PDF binary files for the 4 populous
counties for 2014. These PDF Statement of Votes (SOV) files come from those
county clerks via the SOS office. They are not integrated into the .xls[x]
files because the counties voting machine only output PDF and so that is all
they supply to the SOS. The other 101 counties fill out an online form maintained
by the SOS and that is why they are in the .xls files.